feat:(client) exports to get component or all components
These exports can be used in client side to check if player has any of these categories and check for single components.

Attention!
this only gets player has it in cache , it does not mean player is wearing it
you still must use native to check if player is wearing it.
`IsMetaPedUsingComponent`