fix(client): shop overlay changes
this pr fixes the overlay changes
before this if you remove something it wasnt saving since price was 0

if you had something already with values it would so not charge you to modify and it would not save your changes

it would only save if you change something new.

now we correct this issue but charging any value you modify but not what you initially had.

you you change for color change type change opacity change, go adjust your prices based on that.